101_strands_agent/requirements.txt
strands-agents
strands-agents-tools
strands-agents-tools[rss]
strands-agents-builder
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
feedparser
boto3